Implement router. Add eslint dependencies
