+++ .eslintrc.js
@@ -1,0 +1,25 @@
+module.exports = {
+  'env': {
+    'browser': true,
+    'es6': true,
+  },
+  'extends': [
+    'eslint:recommended',
+    'google',
+  ],
+  'globals': {
+    'Atomics': 'readonly',
+    'SharedArrayBuffer': 'readonly',
+  },
+  'parserOptions': {
+    'ecmaVersion': 2018,
+    'sourceType': 'module',
+  },
+  'parser': 'babel-eslint',
+  'rules': {
+    'linebreak-style': 0,
+    'global-require': 0,
+    'eslint linebreak-style': [0, 'error', 'windows'],
+    'max-len': ['error', {'code': 120}],
+  },
+};
