--- src/webclient/index.js
@@ -1,5 +1,0 @@
-import Application from "./app.js";
-
-const root = document.querySelector("body");
-
-new Application(root, {});
