+++ src/webclient/main/authentication.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>FileHub</title>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-    <link rel="stylesheet" href="../markup_bootstrap/css/style.css">
+    <link rel="stylesheet" href="../../markup_bootstrap/css/style.css">
     <script src="index.js" type="module"></script>
 </head>
 <body>
