+++ src/webclient/main/index.js
@@ -1,0 +1,6 @@
+import Application from './app.js';
+
+const root = document.querySelector('body');
+
+new Application(root, {});
+window.location.hash = '#/';
