+++ src/webclient/tests/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>File Hub tests</title>
+    <meta charset="UTF-8">
+    <link rel="stylesheet" href="../../../node_modules/qunit/qunit/qunit.css">
+</head>
+<body>
+<div id="qunit"></div>
+<div id="qunit-fixture"></div>
+<script src="../../../node_modules/qunit/qunit/qunit.js"></script>
+<script type="module" src="router-test.js"></script>
+</body>
+</html>
